table for fixed slots created. isBreak -> isFixed.
New table created for entering fixed slots.
Role of isBreak changes to isFixed in timeTable table.
For every entry that is isFixed= true, an entry should be
made in fixedEntry table, with a string describing it.
For Lunch slots, the fixedEntry will contain "LUNCH" as string.
Added entry to common.php for the same.

the UI changes for this need to be made